+++ WalletWasabi/CoinJoin/Common/Models/RoundStateResponse4.cs
@@ -1,0 +1,20 @@
+using NBitcoin;
+using Newtonsoft.Json;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using WalletWasabi.JsonConverters;
+
+namespace WalletWasabi.CoinJoin.Common.Models
+{
+	public class RoundStateResponse4 : RoundStateResponseBase
+	{
+		[JsonProperty(ItemConverterType = typeof(PubKeyJsonConverter))]
+		public IEnumerable<PubKey> SignerPubKeys { get; set; }
+
+		public IEnumerable<PublicNonceWithIndex> RPubKeys { get; set; }
+
+		public override int MixLevelCount => SignerPubKeys.Count();
+	}
+}
